+++ core/application.go
@@ -171,7 +171,7 @@ func (app *Application) createCalendarEventForGroups(clientID string, current *m
 			var mappedGroupIDs []string
 			eventID := createdEvent["id"].(string)
 
-			err = app.createCalendarEventForGroupsMembers(clientID, current.OrgID, current.AppID, eventID, newIDs)
+			err = app.createCalendarEventForGroupsMembers(clientID, current.OrgID, current.AppID, eventID, newIDs, nil)
 			if err != nil {
 				return nil, nil, err
 			}
@@ -196,18 +196,29 @@ func (app *Application) createCalendarEventForGroups(clientID string, current *m
 	return nil, nil, nil
 }
 
-func (app *Application) createCalendarEventForGroupsMembers(clientID string, orgID string, appID string, eventID string, groupIDs []string) error {
+func (app *Application) createCalendarEventForGroupsMembers(clientID string, orgID string, appID string, eventID string, groupIDs []string, members []model.ToMember) error {
 	for _, groupID := range groupIDs {
+
 		var userIDs []string
 		memberships, err := app.findGroupMemberships(clientID, model.MembershipFilter{
 			GroupIDs: []string{groupID},
+			Statuses: []string{"admin", "member"},
 		})
 		if err != nil {
 			return err
 		}
 
 		for _, membership := range memberships.Items {
-			userIDs = append(userIDs, membership.UserID)
+			if len(members) > 0 {
+				for _, toMember := range members {
+					if toMember.UserID == membership.UserID {
+						userIDs = append(userIDs, membership.UserID)
+						break
+					}
+				}
+			} else {
+				userIDs = append(userIDs, membership.UserID)
+			}
 		}
 		err = app.calendar.AddPeopleToCalendarEvent(userIDs, eventID, orgID, appID)
 		if err != nil {
@@ -218,14 +229,14 @@ func (app *Application) createCalendarEventForGroupsMembers(clientID string, org
 	return nil
 }
 
-func (app *Application) createCalendarEventSingleGroup(clientID string, current *model.User, event map[string]interface{}, groupID string, members []model.ToMember) (map[string]interface{}, error) {
+func (app *Application) createCalendarEventSingleGroup(clientID string, current *model.User, event map[string]interface{}, groupID string, members []model.ToMember) (map[string]interface{}, []model.ToMember, error) {
 	memberships, err := app.findGroupMemberships(clientID, model.MembershipFilter{
 		GroupIDs: []string{groupID},
 		UserID:   &current.ID,
 		Statuses: []string{"admin"},
 	})
 	if err != nil {
-		return nil, err
+		return nil, nil, err
 	}
 
 	if len(memberships.Items) > 0 {
@@ -236,16 +247,16 @@ func (app *Application) createCalendarEventSingleGroup(clientID string, current 
 
 		createdEvent, err := app.calendar.CreateCalendarEvent(current.ID, event, current.OrgID, current.AppID)
 		if err != nil {
-			return nil, err
+			return nil, nil, err
 		}
 
 		if createdEvent != nil {
 			var mappedGroupIDs []string
 			eventID := createdEvent["id"].(string)
 
-			err = app.createCalendarEventForGroupsMembers(clientID, current.OrgID, current.AppID, eventID, groupIDs)
+			err = app.createCalendarEventForGroupsMembers(clientID, current.OrgID, current.AppID, eventID, groupIDs, members)
 			if err != nil {
-				return nil, err
+				return nil, nil, err
 			}
 
 			for _, groupID := range groupIDs {
@@ -261,11 +272,72 @@ func (app *Application) createCalendarEventSingleGroup(clientID string, current 
 					mappedGroupIDs = append(mappedGroupIDs, mapping.GroupID)
 				}
 			}
-			return createdEvent, nil
+			return createdEvent, members, nil
 		}
 	}
 
-	return nil, nil
+	return nil, nil, nil
+}
+
+func (app *Application) updateCalendarEventSingleGroup(clientID string, current *model.User, event map[string]interface{}, groupID string, members []model.ToMember) (map[string]interface{}, []model.ToMember, error) {
+	memberships, err := app.findGroupMemberships(clientID, model.MembershipFilter{
+		GroupIDs: []string{groupID},
+		UserID:   &current.ID,
+		Statuses: []string{"admin"},
+	})
+	if err != nil {
+		return nil, nil, err
+	}
+
+	if len(memberships.Items) > 0 {
+		var groupIDs []string
+		for _, membership := range memberships.Items {
+			groupIDs = append(groupIDs, membership.GroupID)
+		}
+
+		eventID := event["id"].(string)
+		createdEvent, err := app.calendar.UpdateCalendarEvent(current.ID, eventID, event, current.OrgID, current.AppID)
+		if err != nil {
+			return nil, nil, err
+		}
+
+		if createdEvent != nil {
+			var mappedGroupIDs []string
+			eventID := createdEvent["id"].(string)
+
+			err := app.storage.UpdateEvent(clientID, eventID, groupID, members)
+			if err != nil {
+				return nil, nil, err
+			}
+
+			app.calendar.RemovePeopleFromCalendarEvent([]string{}, eventID, current.OrgID, current.AppID)
+			if err != nil {
+				log.Printf("app.updateCalendarEventSingleGroup() Error cleaning users: %s", err)
+			}
+
+			err = app.createCalendarEventForGroupsMembers(clientID, current.OrgID, current.AppID, eventID, groupIDs, members)
+			if err != nil {
+				return nil, nil, err
+			}
+
+			for _, groupID := range groupIDs {
+				mapping, err := app.storage.CreateEvent(clientID, eventID, groupID, members, &model.Creator{
+					UserID: current.ID,
+					Name:   current.Name,
+					Email:  current.Email,
+				})
+				if err != nil {
+					log.Printf("app.updateCalendarEventSingleGroup() Error create goup mapping: %s", err)
+				}
+				if mapping != nil {
+					mappedGroupIDs = append(mappedGroupIDs, mapping.GroupID)
+				}
+			}
+			return createdEvent, members, nil
+		}
+	}
+
+	return nil, nil, nil
 }
 
 func (app *Application) getGroupCalendarEvents(clientID string, current *model.User, groupID string, filter model.GroupEventFilter) (map[string]interface{}, error) {