+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Implement PUT api/group/{id}/events/v3 API for group calendar events [#394](https://github.com/rokwire/groups-building-block/issues/394)
+
 ## [1.22.0] - 2023-09-26
 ### Added
 - Integrate Calendar BB for group events [#392](https://github.com/rokwire/groups-building-block/issues/392)